+++ lib/puppet/functions/extended_stdlib/sqrt.rb
@@ -7,7 +7,7 @@ Puppet::Functions.create_function(:'extended_stdlib::sqrt') do
     return_type 'Float'
   end
   # @param x The number of which to deteremine the square root.
-  # @return Integer The non-negative square root of the number.
+  # @return Float The non-negative square root of the number.
   # @example Returns square root.
   #   sqrt(4) => 2.0
   #   sqrt(2) => 1.4142135623731